% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_depth.R
\name{func_depth}
\alias{func_depth}
\title{Functional Depth}
\usage{
func_depth(data, times)
}
\arguments{
\item{data}{A matrix where each row is an observation of a time series. Row names should be a unique ID.}

\item{times}{A vector of length equal to the number of columns as data specifying the time points recorded at.}
}
\value{
A numeric vector of depths.
}
\description{
This function calculates the functional depths.
}
