% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_depth_outlier.R
\name{func_depth_outlier}
\alias{func_depth_outlier}
\title{Functional Depth Outliers}
\usage{
func_depth_outlier(
  data,
  times,
  threshold = "beta",
  perc = 0.01,
  B = 10,
  maxiter = 10
)
}
\arguments{
\item{data}{A matrix where each row is an observation of a time series. Row names should be a unique ID.}

\item{times}{A vector of length equal to the number of columns as data specifying the time points recorded at.}

\item{threshold}{The choice of threshold calculation. One of c("beta", "bootstrap"). Default "beta".}

\item{perc}{The percentile of the distribution to use. Default 0.01.}

\item{B}{The number of bootstraps to use. If 0, distribution is fitted to empirical depths. Default 10.}

\item{maxiter}{Maximum number of iterations for outlier detection. Default 10.}
}
\value{
A vector of outliers based on rownames.
}
\description{
This function calculates the functional depth and returns the outliers.
}
\seealso{
\code{\link[=func_depth_threshold]{func_depth_threshold()}}

\code{\link[=func_depth_threshold_beta]{func_depth_threshold_beta()}}
}
