% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_depth_threshold.R
\name{func_depth_threshold}
\alias{func_depth_threshold}
\title{Functional Depth Threshold}
\usage{
func_depth_threshold(data, times, perc = 0.01, B = 1000)
}
\arguments{
\item{data}{A matrix where each row is an observation of a time series.}

\item{times}{A vector of length equal to the number of columns as data specifying the time points recorded at.}

\item{perc}{The percentile of the distribution to use. Default 0.01.}

\item{B}{The number of bootstraps to use. If 0, distribution is fitted to empirical depths. Default 1000.}
}
\value{
A numeric giving the threshold.
}
\description{
This function calculates the threshold for the functional depths in order to
classify outliers using a bootstrap approach.
}
\seealso{
\code{\link[=func_depth_threshold_beta]{func_depth_threshold_beta()}}
}
