% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_depth_threshold_beta.R
\name{func_depth_threshold_beta}
\alias{func_depth_threshold_beta}
\title{Functional Depth Threshold Beta}
\usage{
func_depth_threshold_beta(data, times, perc = 0.01, B = 10)
}
\arguments{
\item{data}{A matrix where each row is an observation of a time series.}

\item{times}{A vector of length equal to the number of columns as data specifying the time points recorded at.}

\item{perc}{The percentile of the distribution to use. Default 0.01.}

\item{B}{The number of bootstraps to use. If 0, distribution is fitted to empirical depths. Default 10.}
}
\value{
A numeric giving the threshold.
}
\description{
This function calculates the threshold for the functional depths in order to
classify outliers by bootstrapping and fitting a beta distribution. This approach is
faster than func_depth_threshold() but makes a Beta distribution assumption for the functional
depths which may not always be appropriate.
}
\seealso{
\code{\link[=func_depth_threshold]{func_depth_threshold()}}
}
