% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_extrapolation.R
\name{func_extrapolation}
\alias{func_extrapolation}
\title{Functional extrapolation}
\usage{
func_extrapolation(data, alpha = 0)
}
\arguments{
\item{data}{A matrix where each row is an observation of a time series. Row names should be a unique ID.}

\item{alpha}{A weighting function to weight between historic and arima forecasts}
}
\value{
A boolean vector.
}
\description{
This function extrapolates partially observed data using historic and arima processes.
}
