% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historic_forecast.R
\name{historic_forecast}
\alias{historic_forecast}
\title{Historic forecast}
\usage{
historic_forecast(d_obs)
}
\arguments{
\item{d_obs}{A matrix where each row is an observation of a time series. Row names should be a unique datetime.}
}
\value{
A numeric matrix
}
\description{
This function computes a historic forecast using linear regression, with days of the week as variables.
}
